refactor ray benchmark helpers
